Run shfmt on mrTest.sh.

+++ bin/mrTest.sh
@@ -51,7 +51,7 @@ done
 # main
 if [ -z "$MAPREDUCE" ]; then
   # shellcheck disable=SC2128
-  if [[ $BASH_SOURCE = */* ]]; then
+  if [[ $BASH_SOURCE == */* ]]; then
     cd -- "${BASH_SOURCE%/*}/" || exit
   fi
   # shellcheck source=/dev/null
@@ -63,25 +63,35 @@ echo "MAP_REDUCE_OUT: $MAP_REDUCE_OUT"
 echo "MAP_REDUCE_JAR: $MAP_REDUCE_JAR"
 
 if [[ $_ACTION == "all" ]] || [[ $_ACTION == "test" ]]; then
-hdfs dfs -rm -r -f "$MAP_REDUCE_IN"
-hdfs dfs -rm -r -f "$MAP_REDUCE_OUT"
+  hdfs dfs -rm -r -f "$MAP_REDUCE_IN"
+  hdfs dfs -rm -r -f "$MAP_REDUCE_OUT"
 
-hdfs dfs -mkdir -p "$MAP_REDUCE_IN"
-hdfs dfs -mkdir -p "$MAP_REDUCE_OUT"
+  hdfs dfs -mkdir -p "$MAP_REDUCE_IN"
+  hdfs dfs -mkdir -p "$MAP_REDUCE_OUT"
 
-hdfs dfs -put -f ./lib/WordCountFile.txt "$MAP_REDUCE_IN"/
-rc=$?; if [[ $rc != 0 ]]; then echo "Input data tranfser failed! exiting"; echo " - MapReduce    - Failed [Input data tranfser failed]" >> "$LOG_PATH"/SummaryReport.txt; exit $rc; fi
+  hdfs dfs -put -f ./lib/WordCountFile.txt "$MAP_REDUCE_IN"/
+  rc=$?
+  if [[ $rc != 0 ]]; then
+    echo "Input data tranfser failed! exiting"
+    echo " - MapReduce    - Failed [Input data tranfser failed]" >>"$LOG_PATH"/SummaryReport.txt
+    exit $rc
+  fi
 
-hdfs dfs -rm -r "$MAP_REDUCE_OUT"
+  hdfs dfs -rm -r "$MAP_REDUCE_OUT"
 
-yarn jar "$MAP_REDUCE_JAR" wordcount "$MAP_REDUCE_IN"/WordCountFile.txt "$MAP_REDUCE_OUT"
-rc=$?; if [[ $rc != 0 ]]; then echo "Mapreduce Job failed! exiting"; echo " - MapReduce    - Failed [Wordcount test failed]" >> "$LOG_PATH"/SummaryReport.txt; exit $rc; fi
+  yarn jar "$MAP_REDUCE_JAR" wordcount "$MAP_REDUCE_IN"/WordCountFile.txt "$MAP_REDUCE_OUT"
+  rc=$?
+  if [[ $rc != 0 ]]; then
+    echo "Mapreduce Job failed! exiting"
+    echo " - MapReduce    - Failed [Wordcount test failed]" >>"$LOG_PATH"/SummaryReport.txt
+    exit $rc
+  fi
 
-echo "******************************************"
-echo "* MapReduce test completed Successfully! *"
-echo "******************************************"
+  echo "******************************************"
+  echo "* MapReduce test completed Successfully! *"
+  echo "******************************************"
 
-echo " - MapReduce    - Passed" >> "$LOG_PATH"/SummaryReport.txt
+  echo " - MapReduce    - Passed" >>"$LOG_PATH"/SummaryReport.txt
 elif [[ $_ACTION == "all" ]] || [[ $_ACTION == "clean" ]]; then
   hdfs dfs -rm -r "$MAP_REDUCE_IN"/WordCountFile.txt
   hdfs dfs -rm -r "$MAP_REDUCE_OUT"
